refactor: move gradient clipping into core utils
